pyproject.toml
[project]
name = "fishqueryplatform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # Django / API
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "drf-yasg>=1.21.10",
    # Server (prod: gunicorn + uvicorn worker)
    "gunicorn",
    "uvicorn[standard]",
    # Observability / HTTP
    "loguru>=0.7.3",
    "requests>=2.32.5",
    # Data
    "psycopg2-binary>=2.9.10",
    # AI / Vector
    "openai>=1.102.0",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = ["ruff>=0.9.1", "pre-commit>=4.0.0", "djangorestframework-stubs>=3.16.6"]

[tool.ruff]
line-length = 88
target-version = "py313"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501", "F403", "F405"]
